gogame: Don't use a useless return but instead a panic(0)
The compiler, for some reason, is not happy if I use log.Panicf() to panic, if I also
would have to return some value. So I used to return something. But sometimes
I want to know which things a function could possibly return and then that extra
return statement is, for my eyes, misleading.