actor: Introduce commands_path
Don't use stdin for commands because I don't open it in my code but the framework
around my code does. So if the writer to the commands becomes disconnected, the stdin closes and
I don't know how to block my goroutine until the stdin becomes available again.
But if it's a file, I can just do open and this will block until that writer connects again.